drivers: spi_nxp_lpsi: fix usage of invalid base address
The "DEVICE_MMIO_NAMED_GET" macro is used before the address space
is mapped, which results in "base" storing NULL (since data->reg_base
is not initialized at this point) instead of the base address of the
IP's address space. Fix this by moving the assignment after the underlying
"device_map()" call.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>